+++ java/src/main/java/se/gewalli/polymorphic_swagger/model/ProductModelV2.java
@@ -5,6 +5,8 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
+
+import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
@@ -69,7 +71,7 @@ public class ProductModelV2 extends ProductModel {
     return this;
   }
 
-  public ProductModelV2 id(Integer id) {
+  public ProductModelV2 id(BigInteger id) {
     super.setId(id);
     return this;
   }
